Update CRUDCollection outputList function
outputList function accepts an (optional) options object with a single "key" property, currently this option is used to switch the collection list from and array of items to an object with the list items as keyed properties - the value of the "key" property is used to select the key values to be used as property names. However the current implementation incorrectly passes the whole listOptions object to the HyperJsonCollection (collection) function instead of just passing listOptions.key, so this feature has a bug.
Additionally the listOptions.key property is later used to supply the items ID attribute name to the autoLink functionality provided by CRUDCollection. These two separate uses for the same option while similar are potentially independent, consider trying to output a list of items as an array while providing the autoLink functionality with the correct item id to generate a working  _self hyperlink.

This patch provides for a second option object property, "listAsKeyedObject" which is truthily tested to allow for all possible use cases to be defined by listOptions. The default behaviour of listOptions (when not supplied) is maintained with original, however the buggy behaviour of listOptions.key usage has been fixed and adjusted, so listOptions.listAsKeyedObject must be truthy for CRUDCollection to generate the items list as an object using item[listOptions.key] values as property names for each item.